review comments by @trustin and @ikhoon
